chore(tasks): mark QA-05 done
