+++ tasks.yml
@@ -1491,7 +1491,7 @@ tasks:
     area: Quality
     dependencies: []
     priority: 4
-    status: pending
+    status: done
     assigned_to: null
     command: null
     actionable_steps:
